Move test and styles in a separated folders for cr suggestion
