+++ src/components/styles/Coins.css
