+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 
-import './index.css';
+import './components/styles/index.css';
 import App from './App';
 import store from './redux/store';
