+++ src/tests/Coin.test.js
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
-import Coin from './Coin';
+import Coin from '../components/Coin';
 
 const mockStore = configureStore([]);
